have a screen to handle products
it is possible now to access a screen to add, modify or delete a product

+++ src/lib/file/provider.ts
@@ -1,12 +1,20 @@
 import ProviderInterface from 'showed/lib/file/service/provider';
 import type Repository from 'showed/lib/file/repository';
 import type { File } from 'showed/lib/file/models/file';
+import fs from 'fs';
 
 export default class Provider implements ProviderInterface {
     constructor(private repository: Repository) {
         this.repository = repository;
     }
     public async deleteFile(id: string): Promise<File> {
+        const file = await this.getFile(id);
+        if (!file) {
+            throw new Error(`File with id ${id} not found`);
+        }
+        if (fs.existsSync(file.filepath)) {
+            fs.unlinkSync(file.filepath);
+        }
         return this.repository.deleteFile(id);
     }
     public async createFile(fileData: { filepath: string }): Promise<File> {
